Extract client camera operations
