+++ custom_components/dreame_lawn_mower/dreame_lawn_mower_client/exceptions.py
@@ -12,3 +12,23 @@ class InvalidValueException(ValueError):
 
 class InvalidActionException(ValueError):
     """ """
+
+
+class DreameLawnMowerError(Exception):
+    """Base reusable client exception."""
+
+
+class DreameLawnMowerAuthError(DreameLawnMowerError):
+    """Login or credential failure."""
+
+
+class DreameLawnMowerConnectionError(DreameLawnMowerError):
+    """Connection or update failure."""
+
+
+class DreameLawnMowerTwoFactorRequiredError(DreameLawnMowerAuthError):
+    """Two-factor authentication is required."""
+
+    def __init__(self, url: str) -> None:
+        super().__init__("Two-factor authentication is required.")
+        self.url = url
